todo funcionando

+++ app/src/main/java/com/example/viewmodelroomjoseph/ui/main/RecyclerActivity.kt
@@ -6,35 +6,36 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import com.example.viewmodelroomjoseph.R
-import com.example.viewmodelroomjoseph.data.HeroRepository
-import com.example.viewmodelroomjoseph.data.HeroRoomDatabase
 import com.example.viewmodelroomjoseph.databinding.RecyclerActivityBinding
 import com.example.viewmodelroomjoseph.domain.Hero
-import com.example.viewmodelroomjoseph.usecases.*
+import com.example.viewmodelroomjoseph.ui.main.adapter.HeroAdapter
+import com.example.viewmodelroomjoseph.ui.main.viewmodel.RecyclerViewModel
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class RecyclerActivity : AppCompatActivity() {
     private lateinit var binding: RecyclerActivityBinding
     private lateinit var heroAdapter: HeroAdapter
-    private val viewModel: MainViewModel by viewModels {
-        MainViewModelFactory(
-            GetHeroes(HeroRepository(HeroRoomDatabase.getDatabase(this).heroDao())),
-            InsertHeroWithSeriesAndComics(HeroRepository(HeroRoomDatabase.getDatabase(this).heroDao())),
-            UpdateHero(HeroRepository(HeroRoomDatabase.getDatabase(this).heroDao())),
-            DeleteHero(HeroRepository(HeroRoomDatabase.getDatabase(this).heroDao())),
-        )
+    private val viewModel: RecyclerViewModel by viewModels()
+
+    override fun onResume() {
+        super.onResume()
+        viewModel.getHeroes()
     }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = RecyclerActivityBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        heroAdapter = HeroAdapter(::getHeroe,::deleteHero,::showData)
+        heroAdapter = HeroAdapter(::deleteHero, ::showData)
         binding.rvHeroes.adapter = heroAdapter
 
-        viewModel.heroes.observe(this,{heroes ->
+        viewModel.heroes.observe(this, { heroes ->
             heroAdapter.submitList(heroes)
 
         })
-        viewModel.getHeroes()
+
 
     }
 
@@ -44,15 +45,23 @@ class RecyclerActivity : AppCompatActivity() {
         bundle.putSerializable(resources.getString(R.string.heroe), hero)
         intent.putExtras(bundle)
         startActivity(intent)
-        finish()
-    }
-    fun deleteHero(hero:Hero){
-        viewModel.deleteHero(hero)
     }
 
-    fun getHeroe(){
-        viewModel.getHeroes()
+    fun deleteHero(hero: Hero) {
+        val dialog = MaterialAlertDialogBuilder(this)
+            .setTitle(resources.getString(R.string.confirmacion))
+            .setMessage(resources.getString(R.string.pregunta))
+            .setNegativeButton(resources.getString(R.string.no)) { view, _ ->
+                view.dismiss()
+            }
+            .setPositiveButton(resources.getString(R.string.yes)) { view, _ ->
+                viewModel.deleteHero(hero)
+
+            }
+            .setCancelable(false)
+            .create()
+
+        dialog.show()
+
     }
-
-
 }
